requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
joblib==1.5.2
numpy==2.3.3
openai==1.109.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.1
pyttsx3==2.99
requests==2.32.5
elevenlabs==1.2.2
scikit-learn==1.7.2
scipy==1.16.2
simpleaudio==1.0.4
sniffio==1.3.1
sounddevice==0.5.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.2.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
